Move NXlog into LoadMuonNexus2
