+++ Framework/Muon/src/LegacyNexusClasses.cpp
@@ -477,89 +477,6 @@ void NXDataSet::getSlab(void *data, int start[], int size[]) {
 //          NXData methods
 //---------------------------------------------------------
 
-NXData::NXData(const NXClass &parent, const std::string &name) : NXMainClass(parent, name) {}
-
-//---------------------------------------------------------
-//          NXLog methods
-//---------------------------------------------------------
-
-/** Creates a property wrapper around the log entry
- * @returns A valid property pointer or NULL
- */
-Kernel::Property *NXLog::createProperty() {
-  NXInfo vinfo = getDataSetInfo("time");
-  if (vinfo.stat == NXstatus::NX_ERROR) {
-    return createSingleValueProperty();
-  } else {
-    return createTimeSeries();
-  }
-}
-
-/** Creates a single value property of the log
- * @returns A pointer to a newly created property wrapped around the log entry
- */
-Kernel::Property *NXLog::createSingleValueProperty() {
-  const std::string valAttr("value");
-  NXInfo vinfo = getDataSetInfo(valAttr);
-  Kernel::Property *prop;
-  NXnumtype nxType = vinfo.type;
-  if (nxType == NXnumtype::FLOAT64) {
-    prop = new Kernel::PropertyWithValue<double>(name(), getDouble(valAttr));
-  } else if (nxType == NXnumtype::INT32) {
-    prop = new Kernel::PropertyWithValue<int>(name(), getInt(valAttr));
-  } else if (nxType == NXnumtype::CHAR) {
-    prop = new Kernel::PropertyWithValue<std::string>(name(), getString(valAttr));
-  } else if (nxType == NXnumtype::UINT8) {
-    NXDataSetTyped<unsigned char> value(*this, valAttr);
-    value.load();
-    bool state = value[0] != 0;
-    prop = new Kernel::PropertyWithValue<bool>(name(), state);
-  } else {
-    prop = nullptr;
-  }
-
-  return prop;
-}
-
-/** createTimeSeries
- * Create a TimeSeries property form the records of the NXLog group. Times are
- * in dataset "time"
- * and the values are in dataset "value"
- * @param start_time :: If the "time" dataset does not have the "start"
- * attribute sets the
- *   start time for the series.
- * @param new_name :: If not empty it is used as the TimeSeries property name
- *   @return The property or NULL
- */
-Kernel::Property *NXLog::createTimeSeries(const std::string &start_time, const std::string &new_name) {
-  const std::string &logName = new_name.empty() ? name() : new_name;
-  NXInfo vinfo = getDataSetInfo("time");
-  if (vinfo.type == NXnumtype::FLOAT64) {
-    NXDouble times(*this, "time");
-    times.openLocal();
-    times.load();
-    std::string units = times.attributes("units");
-    if (units == "minutes") {
-      using std::placeholders::_1;
-      std::transform(times(), times() + times.dim0(), times(), std::bind(std::multiplies<double>(), _1, 60));
-    } else if (!units.empty() && units.substr(0, 6) != "second") {
-      return nullptr;
-    }
-    return parseTimeSeries(logName, times, start_time);
-  } else if (vinfo.type == NXnumtype::FLOAT32) {
-    NXFloat times(*this, "time");
-    times.openLocal();
-    times.load();
-    std::string units = times.attributes("units");
-    if (units == "minutes") {
-      std::for_each(times(), times() + times.dim0(), [](float &val) { val *= 60.0f; });
-    } else if (!units.empty() && units.substr(0, 6) != "second") {
-      return nullptr;
-    }
-    return parseTimeSeries(logName, times, start_time);
-  }
-
-  return nullptr;
-}
+NXData::NXData(const NXClass &parent, const std::string &name) : NXClass(parent, name) {}
 
 } // namespace Mantid::LegacyNexus
